chat tamamlandi, hayvan olusturma database tamalandi

+++ app/src/main/java/com/yasinenessisik/adopt_a_pet/fragments/PetInfoFragment.kt
@@ -1,6 +1,7 @@
 package com.yasinenessisik.adopt_a_pet.fragments
 
 import android.content.ContentValues.TAG
+import android.content.Intent
 import android.os.Bundle
 import android.util.Log
 import android.view.LayoutInflater
@@ -14,6 +15,8 @@ import com.google.firebase.storage.FirebaseStorage
 import com.squareup.picasso.Picasso
 import com.yasinenessisik.adopt_a_pet.R
 import com.yasinenessisik.adopt_a_pet.databinding.FragmentPetInfoFrafmentBinding
+import com.yasinenessisik.adopt_a_pet.model.Post
+import com.yasinenessisik.adopt_a_pet.views.ChatActivity
 
 
 class PetInfoFragment : Fragment() {
@@ -21,6 +24,8 @@ class PetInfoFragment : Fragment() {
     private lateinit var binding: FragmentPetInfoFrafmentBinding
 
     private lateinit var db : FirebaseFirestore
+
+    private lateinit var postobje: Post
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -45,27 +50,56 @@ class PetInfoFragment : Fragment() {
 
             arguments?.let {
 
-                binding.petName.text = PetInfoFragmentArgs.fromBundle(it).pet
-                binding.petBreed.text = PetInfoFragmentArgs.fromBundle(it).petrace
-                Picasso.get().load(PetInfoFragmentArgs.fromBundle(it).imageurl).into(binding.imageView3)
-                var postid  = PetInfoFragmentArgs.fromBundle(it).postid
+                postobje = PetInfoFragmentArgs.fromBundle(it).Post
+
+                Picasso.get().load(postobje.imageurl).into(binding.imageView3)
+
+                binding.petName.text = postobje.petbreed
+                binding.petBreed.text = postobje.petbreed
+                binding.petSpecies.text = postobje.petbreed
+                binding.petAge.text = postobje.petage
+                binding.petCity.text = postobje.petcity
+                binding.petDistrict.text = postobje.petdistrict
+                binding.petGender.text = postobje.petgender
+                binding.petExplanation.text = postobje.petexplanation
+                binding.petUrgency.text = postobje.peturgency.toString()
+
+
+
+
+                var postid  = postobje.docId
 
                 val cameInfo = PetInfoFragmentArgs.fromBundle(it).camefromhome
 
                 if(cameInfo.equals("camefromhome")){
                     binding.delete.visibility = View.INVISIBLE
+                    binding.message.visibility = View.VISIBLE
                 }else {
                     binding.delete.visibility = View.VISIBLE
+                    binding.message.visibility = View.INVISIBLE
                 }
                 binding.delete.setOnClickListener {
-                    db.collection("Post").document(postid)
-                        .delete()
-                        .addOnSuccessListener {
-                            Log.d(TAG, "DocumentSnapshot successfully deleted!")
-                            var action = PetInfoFragmentDirections.actionPetInfoFragmentToMyHomeFragment()
-                            Navigation.findNavController(view).navigate(action)}
-                        .addOnFailureListener { e -> Log.w(TAG, "Error deleting document", e) }
+                    if (postid != null) {
+                        db.collection("Post").document(postid)
+                            .delete()
+                            .addOnSuccessListener {
+                                Log.d(TAG, "DocumentSnapshot successfully deleted!")
+                                var action = PetInfoFragmentDirections.actionPetInfoFragmentToMyHomeFragment()
+                                Navigation.findNavController(view).navigate(action)}
+                            .addOnFailureListener { e -> Log.w(TAG, "Error deleting document", e) }
+                    }
 
+                }
+                binding.message.setOnClickListener {
+                    val intent = Intent(context, ChatActivity::class.java)
+
+
+
+                    intent.putExtra("email",postobje.usermail)
+                    intent.putExtra("uid",postobje.useruid)
+
+
+                    context?.startActivity(intent)
                 }
